+++ __tests__/bookings/resource.test.js
@@ -3,17 +3,20 @@ import Resource from '../../src/bookings/resource'
 const json = {
 	id: 12,
 	name: 'Resource Name',
-	category: 'Cat',
-	status: 'active',
 	quantity: 12,
-	description: 'Some sort of description'
+	description: 'Some sort of description',
+	all_sites: false,
+	site_ids: [
+		1,
+		4
+	]
 }
 
 const resource = new Resource(json);
 
 test('id property', () => expect(resource.id).toBe(12));
 test('name property', () => expect(resource.name).toBe('Resource Name'));
-test('category property', () => expect(resource.category).toBe('Cat'));
-test('status property', () => expect(resource.status).toBe('active'));
 test('quantity property', () => expect(resource.quantity).toBe(12));
 test('description property', () => expect(resource.description).toBe('Some sort of description'));
+test('all_sites property', () => expect(resource.all_sites).toBe(false));
+test('site_ids property', () => expect(resource.site_ids).toStrictEqual([1, 4]));