+++ src/bookings/resource.js
@@ -6,10 +6,10 @@ export default class Resource {
 	constructor(json) {
 		this.id = json.id
 		this.name = json.name
-		this.category = json.category
-		this.status = json.status
 		this.quantity = json.quantity
 		this.description = json.description
+		this.site_ids = json.site_ids
+		this.all_sites = json.all_sites
 	}
 }